+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Alarm scheduler
 
+## Unreleased
+
+* **[NEW]** `CatchThrowableAlarmHandler` wrapper for `AlarmHandler`
+
 ## 0.1.0 (2019-02-15)
 
 * **[NEW]** First tagged version